astconfigparser: Allow escaped semicolons.
When configuring SIP URIs in the pjsip.conf file it is
necessary to escape the semicolon so the parser does not
treat it as a comment. This change allows this to work in
the astconfigparser implementation.

A secondary bug where some data was lost if a configuration
option included a "=" in its value was also fixed.

Change-Id: Ibff2af1c0c4bce28b94360aecec5996f363b4ae8