realtime: Improve test compatibility.
When configuring SIP URIs in the pjsip.conf file it is
necessary to escape the semicolon so the parser does not
treat it as a comment. This change allows this to work in
the astconfigparser implementation.

A secondary bug where some data was lost if a configuration
option included a "=" in its value was also fixed.

Semi-colons are now also encoded according to how the
res_config_odbc module expects. Without this encoding they
were treated as a separator option values.

A bug where sections would be considered equal despite
being different has also been fixed.

Configuration file tests for PJSIP have also been marked
as being for configuration and a new option to runtests.py
has been added to skip a subset of tests based on tags.

Change-Id: Ibff2af1c0c4bce28b94360aecec5996f363b4ae8