realtime_converter: Properly handle multiple values for the same option.
Given the following:
allow=ulaw
allow=g722

The code would currently only insert the first entry for allow into
the database resulting in only ulaw being allowed.

This change fixes this by using the realtime method of separating
multiple values for the same option using ";" allowing both ulaw
and g722 to be allowed.

This is done in the realtime converter itself so this logic will
apply to any multiple values for the same option usage.

Change-Id: I5137be7947a9ad0a2856876222838b8e6cc1feaf