pjsip_session_refresh: Add a test for the PJSIP_SEND_SESSION_REFRESH function
This patch adds a test for the PJSIP_SEND_SESSION_REFRESH dialplan
function. In all cases, the formats on the channel are updated via
PJSIP_MEDIA_OFFER, and then PJSIP_SEND_SESSION_REFRESH is invoked in
some fashion to trigger a session update. The test verifies that the
formats on the channel are updated (or not updated) as expected. This
includes:
 - Update the session using a re-INVITE
 - Update the session using an UPDATE
 - Failure to update the session due to a 488

ASTERISK-26277

Change-Id: Ia3902446f4628ddd52dae1e7b30afd363a902cac