page_predial: Adjust expected bridge channel count.
A recent change to app_confbridge made it so that the announcer channel
is permanently in the bridge. This results in the bridge reporting a
different number of channels than before. This was causing the
page_predial test to fail.

The fix is to account for the change in reported channels when testing.

Change-Id: I220400b2b81d56fb35c144cceb37d6f0f15aa5d9