apptest: Add the ability to specify the channel object from the dialplan
AppTest assumes we have a Local channel pair entering into an
application that we want to control. One half of that Local channel pair
is used by AppTest to send DTMF, stream media, etc. - the other half is
the part that sits in the application under test. In order for AppTest
to know which half of the Local channel is which, it looks at the name
of the Local channel (exten@context) as well as NewExten events to
determine which half has gone into the application under test. This
presents two problems:
(1) Once a Local channel at some exten@context has been 'identified', a
    subsequent Local channel at the same dialplan location will cause
    AppTest to get confused. In the past, different extensions were used
    to control multiple channel pairs; however, there instances where
    the same exten@context has to be used for some tests.
(2) This scheme only works for inbound channels, and cannot handle
    outbound channels. In an outbound channel, the part of a Local
    channel in the application under test is reversed from the inbound
    channel scenario; in fact, we won't get a NewExten event for it at
    all.

This patch solves both of these problems. First, we can now mark a
channel object as being ignored until started. This prevents that
channel object, along with its associated actions, from executing just
because a Local channel exists that would normally match its
'channel-name' field. Once the channel object is started - either
through an action or through the 'start-on-create' property - its event
matchers will be used and considered for action.

Second, we can now identify a Local channel pair from the dialplan.
AppTest will monitor VarSet events and look for one that specifies a
'testuniqueid'. If it matches a channel object's 'testuniqueid' field,
it will associate that Local channel pair with that channel object. This
is completely independent of the application that is under test, and
allows to control both outbound channels as well as other ancillary
channels that might be needed in a complex test scenario.

ASTERISK-25691

Change-Id: I6a2f3da0d0d98f1c3230797b89275a43369e7d8b

Fixup

Change-Id: Ia114421b4a30fe39e77a4ed84be4e04cb8f7e8ef