apps/dial: Add privacy test for call connection
This patch is the first in a series that adds tests for the privacy (P)
and screening (p) options for the Dial application. As such, it consists
of two parts:

(1) It adds a common extensions.conf that will be used by all the tests.
    Since the privacy and screening options include recording a name
    prompt on the caller side and DTMF selection on the callee side, the
    Test Suite AppTest module will be used.

(2) It adds the first test in the series, a test that covers the privacy
    mode where the callee chooses to connect the call. The test does
    basic verification that key sound prompts are played and that the
    call is indeed bridged.

ASTERISK-25691

Change-Id: I1c5143c546c88dc91dee94de8f93445d442d7d81