apps/dial/privacy/privacy: Add a test for the 'banish to VM forever' option
This patch adds a test for the privacy (P) option for the Dial
application, in which the callee decides to banish the caller forever to
VoiceMail. The test verifies that the caller is ejected with the
appropriate DIALSTATUS, and that a subsequent call attempt by the caller
is immediately ejected out of the Dial application without an attempt
being made to reach the callee.

ASTERISK-25691

Change-Id: Idf476a161047bd51384123a6e0c13466a26f6c56