apps/dial/privacy/privacy: Add a test for the 'go away forever' option
This patch adds a test for the privacy (P) option for the Dial application,
in which the callee decides to banish the caller forever to a 'go away
now and forever more' menu. The test verifies that the caller is
ejected with the appropriate DIALSTATUS, and that a subsequent call attempt
by the caller is immediately ejected out of the Dial application without an
attempt being made to reach the callee.

ASTERISK-25691

Change-Id: I103d6a68a7943d8d213d834d571f74a17c501668