apps/dial/privacy/privacy: Add a test for the 'torture' menu option
This patch adds a test for the privacy (P) option for the Dial application,
in which the callee decides to banish the caller forever to a 'torture'
menu. The test verifies that the caller is ejected with the appropriate
DIALSTATUS, and that a subsequent call attempt by the caller is immediately
ejected out of the Dial application without an attempt being made to reach
the callee.

ASTERISK-25691

Change-Id: Ib089169adb213b707707448c60e8025842183a35