apps/dial/privacy/screening: Add a screening test that connects the call
This patch adds a test for the screening (p) option for the Dial application.
Upon receiving the call attempt, the callee selects the option to
connect the call. The test verifies that the call is bridged between the
two parties, along with whether important sound prompts were played to
both parties.

ASTERISK-25691

Change-Id: I79689404aafa59463f86dee71e681e7dd3b0ae65