pjsip/transfers/blind_transfer/off_nominal/transferer_reinvite: Add pause.
This test exercises specific code that sends a BYE
if a re-INVITE is received after a channel has been torn
down. Unfortunately when the re-INVITE is sent in the test
there is no guarantee that this condition has actually
happened yet, in particular on slower systems. This change
introduces a pause so that the channel reaches the proper
state. If it has not yet reached the state a BYE will
still occur but it can take longer than the test expects.

Change-Id: Ifd23662fb6a42a6e9311a816875ba2d188df0851