Get confbridge tests to pass on 13.12.0
A change made to app_confbridge recently made it so that announcements
could be played asynchronously into a confbridge. Prior to that change,
some sounds, such as the conference join sound, would be played twice.
First, the sound would be played to the initiator of the sound (e.g. the
channel joining the conference). Second, the sound would be played into
the entire conference so everyone else could hear it. With the ability
to play sounds asynchronously, we now only play the sound once to the
entire conference where everyone, including the initiator, can hear it.

Asynchronous sounds caused two different types of failures to occur in
tests. The most common failure was simple: tests would check for a sound
to be played to an individual channel, and that no longer happens. The
other type of failure was a failure due to the timing of channels
entering conferences.

For the first type of failure, this was fixed by only looking for the
sound playbacks depending on the version of Asterisk in use. apptest has
been altered to allow a minversion and maxversion to be specified on an
event so that we can decide whether to care about that event or not. The
following tests now make use of a 'maxversion' option:

* confbridge_end_marked
* confbridge_marked
* confbridge_marked_unmarked
* confbridge_triple_lindy
* confbridge_waitmarked_normal

The confbridge_triple_lindy test has one further adjustment. There was a
waitmarked user whose call was previously triggered upon receiving the
event that the join sound had played to a specific channel. That call is
now started when the ConfbridgeJoined event occurs instead.

The confbridge_nominal test was also affected by this failure, but since
it does not use apptest, the solution had to be hard-coded into its
run-test script.

The only test affected by the second type of failure was
confbridge_result. The test made some assumptions about the
synchronicity of the channels that were joining. Specifically, it
assumed that just because a channel called the Confbridge application
first, it would therefore be in the bridge first. This used to be the
case since all channels would block while waiting to get to play
announcements into the bridge. However, now they queue an asynchronous
sound and move on. The order they join the bridge is not predictable. In
order for the test to work in the intended way, the 'Darnell' channel
has to start its call after the 'Alice' channel is confirmed to have
entered the bridge. This is because Alice is endmarked and Darnell
is marked. The test assumes that when Darnell leaves the bridge, Alice
will be kicked out as well. However, in test runs, Darnell would enter
and leave the bridge before Alice ever entered.

Change-Id: I4bcaaf5b535c34f636d0ccc66a6772db7a062de2