Testsuite - codec_opus: Add encode/decode test
Tests encoding and decoding of opus in Asterisk. For each test a call is
originated into app record and either a tone or audio file (containing a tone)
is played back. Once done the recorded file is analyzed to make sure that the
original tone matches the recorded tone.

ASTERISK-26409 #close

Change-Id: I961d483d87b5d1439e20171cee859c1e1606c824