pjsip/attribute_passthrough/speex_h263_h264: Disable direct media.
The fixing of ASTERISK-26423 has resulted in direct media being
attempted for cases where it would not have previously occurred.
This is because the RTP native bridge check was getting all
configured codecs (not negotiated). If video was configured then
no direct media would be attempted, even if video was not
negotiated. This is no longer the case so direct_media needs
to be explicitly disabled.

ASTERISK-26423

Change-Id: Iacdc75c43978cd1ffce5f01c792bb73805ed2fe3