Allow apptest to work with master.
Master tests that use the app_test module were failing under the master
branch. This is due to the commit that added version checking for
individual events. The code attempted to set arbitrary version
boundaries of "0.0.0" and "999.999.999" if no version was specified in
the configuration. The problem is that when doing version comparisons,
"GIT-master" gets changed into sys.max_int. This means that it will
always be greater than any maxversion you specify.

The fix here is to no longer attempt to use arbitrary values when
versions are not specified. If a version is not specified, we do no
comparison and just assume the test is runnable with the current running
version.

Change-Id: Id30e7c3a57cb720d070cc5677aa9960bd40a8ada