Testsuite: avoid accidental use of alternate tmp
The alternate /var/tmp directory for asterisk-testsuite
temporary files will be autoamtically selected- when it
has more blocks available then /tmp. This can happen in
CentOS 7 where /tmp is mounted on a smaller tmpfs.

However if /tmp and /var/tmp are on the same filesystem
and the available space changes slightly between calls,
the testsuite could mistakenly select /var/tmp where it
was not necessary.  By changing the comparison function
to treat small differences in available space the same,
this problem is avoided.

ASTERISK-26602 #close

Change-Id: Ia1bc3ded873e917960478f31c257c93cd28fb4aa