test_case.py: Ignore reactor timeout if already stopping.
Some tests have a hook to fail the test if the reactor times out during a
test as a fail safe.  However, if the reactor timeout happens while the
test is waiting for Asterisk to shutdown then we should simply log the
event but otherwise ignore it.  The test has already completed and it is
waiting for cleanup to happen because of the CLI "core stop gracefully"
command.  That shutdown command can take awhile to complete due to delays
required for things to safely settle down for cleanup.

Increasing the shutdown timeout gives the reactor timeout more time to
fire and potentially erroneously fail a test.  At this point in a test we
are already shutting down Asterisk.  The shutdown timeout will forcibly
kill Asterisk when it fires.

ASTERISK-26527

Change-Id: Ida565f6151d396821db8770382b6624aec16558e