tests/rest_api/events/variables: Stop test on StasisEnd.
There was a race condition in this test where a StasisEnd
would not be received as Asterisk, and the test, were being
terminated during that time as a result of the channel hanging
up.

This change makes the test (and Asterisk) stop as a result of
the test receiving StasisEnd instead.

ASTERISK-26492

Change-Id: Ie5faa3a2aea1280b6c89d570a592f9dfdedc40b5