Adding tests that test forward error correction in codec_opus
When FEC is enabled and negotiated in codec_opus dropped frames can be built
from a subsequently received frame. This series of tests establish a call
between two endpoints. Once the call is 'up' RTP packets are 'played' back
from a pcap file that contains roughly a minute of opus audio. Several packets
have been deleted in order to simulate dropped packets. The other end of the
call is monitored to make sure packets are received in the correct order and
with the appropriate timestamp/interval. If FEC is working properly there should
be no gaps in the sequence numbers of the received packets. Also the difference
between the timestamps should not vary.

These tests also make sure FEC works with and/or without jitter buffers enabled.

Change-Id: Ia01a0432cc23071836d594c7c3dc46709c9880bf