pjsip/registration/.../regcontext:  Use better start trigger
The regcontext test was using the CONTACT_ADDED TEST_FRAMEWORK
event as the signal to check the sipregs context for the registered
extension but that event is actually generated BEFORE the extension
is added to the context.  Depending on system capacity, debug/on/off,
realtime, etc., the check therefore could return an empty dialplan
list.

Changed the trigger event to PeerStatus/Reachable which is only
generated after the extension add completes.  This also removes
the test's dependency on TEST_FRAMEWORK.

Change-Id: I0d4f975489cd6789b10bc1e5a384403b17b33aa1