tests/channels/pjsip/subscriptions/ast_restart: New test.
Test that subscriptions survive an Asterisk restart.

This is the first pluggable test that causes Asterisk to be restarted as
part of the test.  As such, the AMI EventActionModule, TestCase, and the
SIPpTestCase needed some new code to handle an Asterisk restart.

* Added 'allow-ami-reconnects' boolean option to inform the TestCase base
class that AMI must be able to reconnect after Asterisk is restarted
during the test.  The option defaults to False.  The option is only needed
because old versions of starpy did not support reconnecting the AMI
session if the connection is broken for some reason.  The option basically
ensures that the test will fail with a meaningful error message if the
version of starpy does not support reconnects.

* Added support for the ami-start event of the AMI EventActionModule to
distinguish which Asterisk instance AMI got started on.  The optional
instance id defaults to '0'.

ami-config:
    ami-start:
        id: '1'
    ami-actions:

* Added the new ami-restart event of the AMI EventActionModule to trigger
actions.  The ami-restart event happens when the AMI connection is
reconnected and Asterisk is fully booted.  You can optionally specify
which Asterisk instance (defaults to '0') the event applies.  You can
optionally keep track of the 'count' of restart events required for the
test to pass.  You can optionally trigger the action only on the specified
event count.  Otherwise, the actions are triggered on every event.

ami-config:
    ami-restart:
        id: '1'
        count: '>1'
        trigger-on-count: True
    ami-actions:

Change-Id: I545a7921f5d201b26140c00156777c4c0c9e5dd7