tests/channels/pjsip/subscriptions/ast_restart: Fix race condition.
The test is failing on fast machines because they restart and fully boot
before the testsuite reconnects the AMI connection.  As a result the
expected second SUBSCRIPTION_ESTABLISHED event is never seen.

* Instead of ping-ponging back and forth between Asterisk and the SIPp
scenario, change the test to be more SIPp scenario driven.

* Enable PJSIP message debugging in the config file to ensure that logging
messages are enabled early enough to see the subscription recreation
NOTIFY message on fast machines.

Change-Id: Ie79c5ca9954287c1dda42a24305ccc8bdd0a6e07