rtcp: Add PJSIP tests for RTCP-MUX support.
This adds tests for the various RTCP-MUX configuration
possibilities and confirms that the flow of the underlying
RTCP matches what is expected for the specific case. This
includes tests which use ICE to determine the route to use.

The tests work by limiting the ports in use by the
res_rtp_asterisk module such that they are always
predictable. This allows us to determine using the
RTCPReceived AMI events the source and target ports
for RTCP traffic based on the various configurations.

ASTERISK-26732

Change-Id: I8d5624a83a7bf9d153196c4fbd402e3f7df4a599