Add a nominal/off-nominal test for identify by header
This patch adds a test that covers both nominal and off-nominal matching
by header. In the nominal scenario, an inbound request is matched
correctly using a SIP header and a matching token. In the off-nominal
scenario, an inbound request is rejected due to having the correct
header but incorrect token value.

ASTERISK-26863

Change-Id: Idd02adba1ea97f146d5cfc37fe5f5d1a0665f45f