Remove hard coded tcp ports in SIPP tcpauthlimit tests
If a test attempts to use a local TCP port that was just used by
a previous test, chances are that that port will still be in
TIME_WAIT state and the test will fail.

Removed the hard coded local ports in the sipp_client_scanrio
and tcp_client_scenario tests.

Re-enabled the tcp_client_scenario test.

Change-Id: Ifa93373193b0933a99cad4fd4741190652656745