testsuite:  Move pjsip/transfers/../callee_local_anonymous to headers
We've determined that anonymizing the From header when doing a
callee transfer re-invite is not correct.  There are no other
anonymized From tests though so this one was moved to the "headers"
directory and converted to do just a basic call with
callerid_privacy=prohib set.  The UAS (Bob) verifies that the From
header on the INVITE is anonymized.

A second test was also added to confirm that when
callerid_privacy=prohib is NOT set, the From header isn't
anonymized.

Change-Id: I5b67ff0b7b1909b1db2edcd2f982a6a057c8c12e