tests/rtp/strict_rtp: Add test for strict RTP support.
This adds a test that sets up a call between two Asterisk
instances with strict RTP enabled and symmetric RTP enabled.
Once media has started flowing a flood of minimal RTP packets
are sent to each instance. If any traffic is received on the
port that the RTP packets were sent from the test fails as
strict RTP protection did not discard the packets.

ASTERISK-27013

Change-Id: Id24462bed062424295d32618658d94ef4209a934