testsuite:  Add delays to pjsip/ami registration tests
The sipp scenarios for show_registration_inbound_contact_statuses
and show_registrations_inbound sent a REGISTER and waited for
an OK but then ended sometimes causing the tests to fail because
there wasn't enough time to receive the AMI events they're
testing for.

Added a 3 second delay after the receipt of the OK to keep the
scenario alive long enough to capture the events.

Change-Id: I29821cc81509a54c1a234507ee22c86dbceac436