pjsip/rtp/asymmetric_rtp_codec: Add tests for local format behavior.
This adds two tests which confirm the behavior of the
'asymmetric_rtp_codec' option for local formats.

One test confirms that when set to 'no' the resulting formats
on the channel contains only one format.

The other test confirms that when set to 'yes' the resulting
formats on the channel contain all negotiated formats.

ASTERISK-26996

Change-Id: I52c84ec317d8b4ead82d5974227a3d5c15f83808