Test receiving an early BYE during an attended transfer
Asterisk used to fail the transfer and drop the call if an early BYE (sometime
just after the 202 accepted, but before the Notify 200 OK frag) was received.
Added a test that makes sure the transfer does not fail when an early BYE is
received.

This patch modified the testsuite's pjsua phones wrapper to allow configuring
of when to send the transferer's BYE during the transfer process. A new option,
'hangup-reason', has been added that specifies when to send the BYE. Allowed
values are:

  accepted - Sent right after the refer is accepted
  trying - Sent right after receiving a sip frag 100 Trying
  ok - Sent right after receiving a sip frag 200 OK
  final - Sent upon receiving the final message

A test was also added that sends a BYE right after the REFER is accepted. As
noted in test-config.yaml this test is a direct copy of the 'same_stasis_app'
test except for when the BYE is sent.

ASTERISK-27053

Change-Id: Ifbf1d6e8f91dfc00128a2ccd7eaff6402d170650