pjsip: Correct attribute passthrough tests.
While testing multistream support in PJSIP it uncovered
a few issues with the attribute passthrough tests.

The speex test was not configured to accept both ulaw
and speex. As a result in multistream a reinvite occurred
for the calling leg to switch it to only speex. This change
makes both ulaw and speex accepted so that no reinvite occurs.

The h263 and h264 tests did not answer their INVITEs with
valid SDP. They answered with only a single video stream
when they should answer with both an audio and video. In
multistream this caused the calling leg to get a reinvite
declining the audio stream. PJSIP helps in this area by
correcting the received SDP so it is proper.

Change-Id: I966e7c53831e0ba925e3c8fc500953b9c46a4b1b