single_resource_multiple_changes:  Refactor to remove custom python
This test (and many other rls tests) had a custom python thing
that proxied the sip stream between the test and asterisk and
resulted in 3 state machines running at the same time.  All for
the sole purpose of testing that when rls notifies are batched,
the test only gets the last notify.  It was unstable in high
stress environments.

Replaced with a few extra lines of xml in the scenario file and
4 new ami-event/ami-action pairs in yaml.

Change-Id: Ibfb23af16898cd0780f46deca03abb2815c71c96