testsuite: Add min/max version lists to test-object and modules.
Make testsuite accept the minversion/maxversion syntax for "test-object"
and "modules".  Before it only accepted a single version string.  Now it
will accept a single version string or a list of version strings.

test-modules:
    test-object:
        -
            maxversion: ['13.18.0', '14.7.0', '15.1.0']
            config-section: test-object-config-v12
            typename: 'sipp.SIPpTestCase'
        -
            minversion: ['13.18.0', '14.7.0', '15.1.0']
            config-section: test-object-config-v13.18.0
            typename: 'sipp.SIPpTestCase'
    modules:
        -
            maxversion: ['13.18.0', '14.7.0', '15.1.0']
            config-section: 'ami-config-v12'
            typename: 'ami.AMIEventModule'
        -
            minversion: ['13.18.0', '14.7.0', '15.1.0']
            config-section: 'ami-config-v13.18.0'
            typename: 'ami.AMIEventModule'

Also fixed a bug in the "properties: maxversion" handling of version
lists.  We were cutting off running the test only if the Asterisk version
was greater than or equal to all listed versions.  It is now handled
similarly to minversion.

Change-Id: Ibd896104af60bdab720e14e3b5bd03aafc8f2167