pjsip/message:  Add test for passing message through confbridge
Alice and Bob call a confbridge.  Alice waits a few seconds to make sure
Bob is in the bridge then sends an in-dialog MESSAGE with a text/html
body.  Bob gets the MESSAGE and verifies that Content-Type was preserved
and that the body is intact.

Change-Id: Icde9beeb210ba42be5d9bc391f0e3ab6c11c3479
(cherry picked from commit 3fd342d085a38c8ff86da415c87ad230a2583ed1)