channels/pjsip/qualify/call_unreachable: Trigger on a more reliable event
On my test box, the test occasionally never saw the ContactStatus event on
the AMI connection.  The ContactStatus event is only sent once per status
change and likely got "sent" before the test established the AMI
connection.  The test was using that event to trigger sending the
Originate action and thus the test fails.

* Switch to using the TEST_FRAMEWORK test event 'TestEvent' for the
AOR_CONTACT_QUALIFY_RESULT state.  This event is sent periodically and is
a more reliable trigger event to initiate the Originate action.

* Added some more test failure checks.

Change-Id: Id9500c988b4f6cac44783fcd7663dc2cbef4add5