Add support for delayed shutdown of Asterisk when checking for leaks.
Add memcheck-delay-stop option to test-config.yaml in the test-object
config section.  This causes a pause during shutdown if Asterisk is run
under valgrind or has REF_DEBUG enabled.  This delay is not meant to
compensate for slowness caused by memory debugging, but to allow SIP
timers to clean resources before shutdown of Asterisk is initiated.

If set this options forces reactor-timeout to be at least 3 seconds
longer.  This is only done to stop the testsuite from printing a reactor
timeout warning for tests with a large memcheck-delay-stop value.

I had considered setting a non-zero default for memcheck-delay-stop but
the majority of tests do not require any delay.  Since the testsuite
contains nearly 1000 tests using a 7 second default delay would add
nearly 2 hours to the total runtime.  I estimate that only 20% of the
tests require any delay.  The amount of delay needed varies based on
which timers we need to wait for: timer_b, session expiration, defered
session termination.  The minimums for these timers range from 6.4 - 90
seconds.  In most cases using the minimum delay will require modification
of the test.  For example reconfiguring timer_t1 and timer_b in some basic
PJSIP call tests or reducing the 'Expires:' header in SIP SUBSCRIBE tests.

As an example of usage tests/channels/pjsip/accountcode has been
updated.  The timer_t1 and timer_b have been decreased to minimize the
delay needed.

ASTERISK-27306

Change-Id: I67ec89e4a116526ad34796a233e15fec794cf3e1