pjsip: Add tests for 'ip' option of 'identify_by' on endpoint.
This change adds two tests:

The first test sends various calls to Asterisk that initially match
based on the username portion of the From header. The endpoint
is configured to not allow it to be matched based on username.
Afterwards matching based on IP address occurs to a valid
endpoint.

The second test sends various calls to Asterisk that initially match
based on the source IP address of the INVITE. The endpoint is
configured to not allow it to be matched based on this. Afterwards
matching based on username occurs to a valid endpoint.

In both tests if calls work then the right endpoint has been
matched. If calls fail then the endpoint was matched incorrectly.

ASTERISK-27206

Change-Id: Ieb4e504b0063c723f6cdafd0f78281e33027f7ac