channels/pjsip/transport/reuse: Tests for reuse of TCP transports.
These tests check to see if the TCP transport that was used to register
the endpoint is reused.  The tests depend upon the rewrite_contact option
because the REGISTERed contact points to a nonexistent location.  If the
transport isn't reused then any SIP request sent by Asterisk will fail and
cause the test to fail.

Change-Id: Id89af15fb3e7a63bd58d6115d923cb30a07d99ae