res_pjsip: qualify/auth test failed with ref leaks when using REF_DEBUG
When REF_DEBUG was enabled the pjsip/qualify/auth would fail due to the
req_wrapper reference being leaked. This happened because a second QUALIFY
was being sent out while the test was ending (Asterisk was being shutdown).
So the send request callback was never called, due to Asterisk disappearing,
thus not releasing the memory.

For now the solution is to just increase the qualify frequency so only a
single qualify is sent for the test.

ASTERISK-27358 #close

Change-Id: If597268978ea79852e9e339a2a56a8ac7d2f63dd