res_pjsip - chan_is_avail fails if running with sipp version < 3.5.0
sipp 3.5.0 made it so the "-aa" option responds to OPTION messages. However,
versions prior to that would not. The test config incorrectly listed the
sipp an older version for the dependency.

This patch bumps the sipp dependency version to the appropriate value, so now
anyone running this test will skip the test instead of failing.

Change-Id: I5147b726cc1a07aa7293e579bbb0e6041d1c1b66