Update contact status related test with NonQualified nomenclature
A few tests were checking specifically for the "Created" status. This status
has now changed to the more descriptive "NonQualified" status. This patch
updates the relevant tests.

Also the rest-api subscribe-all test was not setting the qualify_frequency, so
the endpoint never became "Reachable". This patch also updates that test to
specify a qualify_frequency.

ASTERISK-27467 #close

Change-Id: Icc6c41f42f9fd465c414b4f3a9eeda2ea15536fa