rest_api/applications/subscribe-all: Add delay to shutdown.
This should hopefully stop the test from randomly failing.  I ran the
test 100 times each with different delays.  The test failed 15 times
when run without any delay, 4 times when run with 2 second delay, no
failures when run with a 3 second delay.

I realize this is not ideal but until someone finds a better solution
this is better than nothing.

Change-Id: Ic0faccc5443a3ca800d887e3169a3336634d60c7