test_case.py: Conflict dependencies not being processed
When the testsuite installs config files it checks the 'conflicts.txt' file and
makes sure any conflicting modules do not load together. This feature was no
longer being used due to a wrongly passed in parameter to the install_configs
method. Instead of passing in the dependency list it passed in the config
object itself.

This patch makes it so the dependencies are once again passed into the method.

Change-Id: Icdac9a9b91d6c2a682f9e2124296fdcfbe3b96a0