confbridge_talking: Add test that checks talking status with hold
If a channel in a softmix bridge is put on hold then it's talking status is
cleared and an AMI event is raised. This test checks for that event and whether
or not the talking status has been set to "on" or "off" when the channel is
put on hold.

ASTERISK-27755

Change-Id: I99e343439c5a43f178fc466aa2978f6dfd1ab73a