sipp, test_suite_utils: Default media port to an unused port
SIPp correctly chooses an available port for audio, but unfortunately it then
attempts to bind to the audio port + n for things like rtcp and video without
first checking if those other ports are unused:

https://github.com/SIPp/sipp/issues/276

This patch makes it so all ports needed by the scenario are available. It does
this by retrieving and checking for an unused port plus 'n' ports from the OS.
If the ports are available then the primary port is passed to the scenario using
the '-mp' option.

Change-Id: I3da461123afc30e1f5ca12e65d289eaa42d6de00