pjsip_channel: Use specified media port
This test took advantage of the fact that SIPp defaults the media port to 6000,
which it then checked against for the test. However, a recent patch made it so
tests using the SIPp 'media_port' option would get an available random port
instead.

This patch specifies the media_port for SIPp to use and makes sure it is the
same one checked by the test. A number was chosen which hopefully makes port
collisions highly unlikely.

Change-Id: Id7f82d8245401659d269e6ba084c2ad05aee068d