utils_socket: range of random ports searched only once
If a range (greater than 1) of random ports was asked for, and if a port in
range was not available an exception was thrown stating the port in range
was not free.

This patch makes it so the exception is caught and a new random port range is
tried again (up to "attempts" tries) before quitting.

Change-Id: I96affa9df38acf569f5d3c0965ac7b78d865455d