utils_socket: don't reuse port variable
The port parameter variable was being written over and reused in the 'get_range'
random port retrieval section. If a second attempt needed to be made to get a
free port then it would end up falling into the first case (port != 0) and
cause a concatentation error since port was now a list.

This patch makes it so the second case uses a completely new variable name.

Change-Id: I9e540eb5a1edddb9f6dd9ae625a3cdc255e60820