pjsip/statsd/contacts: Sporadically failing due to unexpected messages
After the pjsip qualify rewrite the contacts statsd test failed fairly
regularly because it was now receiving two messages that it did not receive
before the rewrite. Namely, the messages that occur on shutdown. However,
due to some current shutdown handling in Asterisk these final messages still
may or may not be received by the test.

Unfortunately, the way the statsd tests were written did not allow for optional
messages. Really it only allowed for a strict result set that was not too
configurable.

This patch alleviates the brittleness, and fixes the test, by creating
suitable condition based message handling and matching routines that the test
can now use for optional messages received.

An attempt was made to generically write the message matching routines in hopes
that other future event listeners and handlers could take advantage of it.

Change-Id: Iaae769c7a4fe2dcac4865eb7dc4e5b6a1b25900b