apps/dial: Convert to use chan_pjsip
The sip_to_pjsip script took care of the sip.conf to pjsip.conf
conversion but there were other issues.

The regex in sipp.xml files to capture the to tag was wrong.
"s" now gets returned as ConnectedLineNum in many cases.
"NO ANSWER" now gets returned as a dial status when the remote
end doesn't answer instead of "BUSY"

This commit is dependent on the mainline asterisk commit...
"chan_pjsip:  Register for "BEFORE_MEDIA" responses"

Change-Id: If176aff9ca48a9e7037514c202b83b270eb52537