res_pjsip_pubsub/exten_state: state changes sometimes initiated too soon
There is a race condition in the test where the state changes could be triggered
before the subscription was finalized. If this happened then the test would time
out and fail due to not receiving the expected NOTIFY's.

This patch ensures that the test does not trigger the state changes until the
subscription has been fully established.

ASTERISK-27765 #close

Change-Id: I3ae529d475a8b148eb06b536d9f37d4778b40c52