endpoints/statsd: Update test to use new condition based matching
The statsd tests were written in a way that did not allow for flexible message
matching. A previous patch added condition based message handling and matching
routines that allow greater flexibility in matching messages.

This patch updates the endpoints/statsd test to take advantage of that. It also
removes an unneeded dependency.

Change-Id: I4c3b5e1efe063b047d41f8c7f4c4674324af2650